fix: separate query/command handler files into dedicated files (closes #103)
