+++ backend/src/Chickquita.Application/Features/FlockHistory/Commands/UpdateFlockHistoryNotesCommandHandler.cs
@@ -1,0 +1,70 @@
+using AutoMapper;
+using Chickquita.Application.DTOs;
+using Chickquita.Application.Interfaces;
+using Chickquita.Domain.Common;
+using FluentValidation;
+using MediatR;
+
+namespace Chickquita.Application.Features.FlockHistory.Commands;
+
+/// <summary>
+/// Handler for UpdateFlockHistoryNotesCommand.
+/// </summary>
+public sealed class UpdateFlockHistoryNotesCommandHandler
+    : IRequestHandler<UpdateFlockHistoryNotesCommand, Result<FlockHistoryDto>>
+{
+    private readonly IFlockHistoryRepository _repository;
+    private readonly IMapper _mapper;
+    private readonly IValidator<UpdateFlockHistoryNotesCommand> _validator;
+    private readonly IUnitOfWork _unitOfWork;
+
+    public UpdateFlockHistoryNotesCommandHandler(
+        IFlockHistoryRepository repository,
+        IMapper mapper,
+        IValidator<UpdateFlockHistoryNotesCommand> validator,
+        IUnitOfWork unitOfWork)
+    {
+        _repository = repository;
+        _mapper = mapper;
+        _validator = validator;
+        _unitOfWork = unitOfWork ?? throw new ArgumentNullException(nameof(unitOfWork));
+    }
+
+    public async Task<Result<FlockHistoryDto>> Handle(
+        UpdateFlockHistoryNotesCommand request,
+        CancellationToken cancellationToken)
+    {
+        // Validate the command
+        var validationResult = await _validator.ValidateAsync(request, cancellationToken);
+        if (!validationResult.IsValid)
+        {
+            var errors = validationResult.Errors
+                .Select(e => e.ErrorMessage)
+                .ToList();
+
+            return Result<FlockHistoryDto>.Failure(
+                Error.Validation(string.Join("; ", errors)));
+        }
+
+        // Find the history entry (tenant isolation handled by RLS and global filters)
+        var historyEntry = await _repository.GetByIdAsync(request.HistoryId);
+
+        if (historyEntry == null)
+        {
+            return Result<FlockHistoryDto>.Failure(
+                Error.NotFound("Flock history entry not found"));
+        }
+
+        // Update the notes using the domain method
+        historyEntry.UpdateNotes(request.Notes);
+
+        // Save changes
+        await _repository.UpdateAsync(historyEntry);
+        await _unitOfWork.SaveChangesAsync(cancellationToken);
+
+        // Map to DTO and return
+        var dto = _mapper.Map<FlockHistoryDto>(historyEntry);
+
+        return Result<FlockHistoryDto>.Success(dto);
+    }
+}
